Allow 3rd parties to post new slack methods without changes
